requirements.txt
pygame==2.6.1
svg.path==6.3
